Docs tweaks

+++ tests/test_bake_project.py
@@ -1,16 +1,9 @@
 from contextlib import contextmanager
 import shlex
 import os
-import sys
 import subprocess
-import datetime
 from cookiecutter.utils import rmtree
-
-from click.testing import CliRunner
-
-import importlib
-import pytest
-
+from pytest_cookies.plugin import Cookies
 
 @contextmanager
 def inside_dir(dirpath):
@@ -27,17 +20,18 @@ def inside_dir(dirpath):
 
 
 @contextmanager
-def bake_in_temp_dir(cookies, *args, **kwargs):
+def bake_in_temp_dir(cookies: Cookies, *args, **kwargs):
     """
     Delete the temporal directory that is created when executing the tests
-    :param cookies: pytest_cookies.Cookies,
+    :param cookies:
         cookie to be baked and its temporal files will be removed
     """
     result = cookies.bake(*args, **kwargs)
     try:
         yield result
     finally:
-        rmtree(str(result.project))
+        if result.project:
+            rmtree(str(result.project))
 
 
 def run_inside_dir(command, dirpath):
@@ -51,11 +45,13 @@ def run_inside_dir(command, dirpath):
 
 
 def test_bake_and_run_tests(cookies):
-    my_extra_context={'full_name': 'fakename',
-                    'email': 'email@email.com',
-                    'project_name': 'myproj',
-                    'use_github': 'y'}
+    my_extra_context={
+        'full_name': 'fakename',
+        'email': 'email@email.com',
+        'project_name': 'myproj'
+    }
     with bake_in_temp_dir(cookies, extra_context=my_extra_context) as result:
+        assert not result.exception
         assert result.project.isdir()
         install = run_inside_dir('pipenv install --dev', str(result.project))
         assert install.returncode == 0
@@ -63,8 +59,9 @@ def test_bake_and_run_tests(cookies):
         assert test.returncode == 1
         out = test.stdout.decode()
         print(out)
-        assert "4 failed" in out
+        assert "5 failed" in out
         assert "Please change ./README.rst" in out
         assert "Please change description in ./setup.cfg" in out
-        assert "Please change the documentation in docs/index.rst" in out
-        assert "Please change the documentation in docs/reference/api.rst" in out
+        assert "Please change ./docs/reference/api.rst" in out
+        assert "Please delete ./docs/how-to/accomplish-a-task.rst" in out
+        assert "Please delete ./docs/explanations/why-is-something-so.rst" in out
